+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 All notable changes to Aelys, roughly grouped by version. I don't always tag releases perfectly, so this is reconstructed from git history
 
-## 0.19.x - Array, Vec, and @inline decorator implementations
+## 0.19.x - Array, Vec, and @inline decorator, and bugs fixes
+
+**0.19.8-a** (not a "real" update, I just had to make a new git tag for something to work, this contains some bug fixes)
+- `-ae.trusted=true` flag wasn't working
+- FFI string support
+- vec methods on dynamic-typed parameters
+- Some std fixes
+
+**0.19.7-a**
+- Fixed vec methods on dynamic-typed parameters
+- FFI string support
 
 **0.19.6-a**
+- Fixed string equality across modules and Vec methods on Dynamic-typed parameters
+- Added [VSCode](docs/syntax-support/vscode/README.md) syntax highlighting
 - UDP support by @LeKebabiste, see [Acknowledgements](ACKNOWLEDGEMENTS.md)
 - Remove aelys- prefix from crate directory names